+++ data_engineering/utils/pipeline_io.py
@@ -4,7 +4,7 @@ import os
 
 
 def get_para() -> dict:
-    file_path = os.path.join(os.path.abspath(""), "..", "conf/parameters.yaml")
+    file_path = os.path.join(os.path.abspath(""), "conf/parameters.yaml")
     print(f"path to the config file: {file_path}")
     with open(file_path) as file:
         paras = yaml.load(file, Loader=yaml.FullLoader)
